Make script more robust

+++ create-symbolic-links.sh
@@ -1,9 +1,20 @@
 #!/bin/bash
 
-rm ~/.bashrc
-#rm ~/.vimrc
-#rm ~/.spacemacs
-rm ~/.zshrc
+if [ -e ~/.bashrc]; do
+    rm ~/.bashrc
+done
+
+if [ -e ~/.vimrc]; do
+    rm ~/.vimrc
+done
+
+if [ -e ~/.zshrc]; do
+    rm ~/.zshrc
+done
+
+if [ -e ~/.spacemacs]; do
+    rm ~/.spacemacs
+done
 
 ln -s ~/git-repos/dotfiles/.bsahrc ~/.bashrc
 ln -s ~/git-repos/dotfiles/.vimrc ~/.vimrc
